feat: separacao da aplicacao em modulos
ref #1
