+++ .dockerignore
@@ -8,3 +8,4 @@ gitlab
 LICENSE
 *.md
 Dockerfile
+docs
